Enable input fields listed in URL address
When entering the URL which already has a query string, if there are
no elements satisfying the query, all the input fields become disabled.
This commit enables all the fields which are listed in the query string
even if none of the elements meets the specified filters.
Also, "vcs.xml" is a new file created by PyCharm which is also added.