Fix in_right_unit function and rewrite it
This commit fixes a comparison when defining the correct unit for a
value. Before this commit, MinValue and MaxValue was either increased or
decreased by 1e-10 depending on the MinIsIncluded and MaxIsIncluded
being False respectively, and only one non-strict inequality was used.
This lead to incorrect unit conversions.

For example, if the parameter value was 1e9 and MaxIsIncluded was False,
then MaxValue would be decreased by 1e-10. However, (1e9 <= 1e9 - 1e-10)
returns True and causes a wrong unit to be appended to the parameter
value. Also, "add_unit" function was created to avoid repeating.