Fix input field values when entering URL directly
This commit adds URLValueToInput and appendUnit functions, which process
range values from URL (when directly entering it) and fills the input
fields with the correct representation including unit if it is needed.

Before this commit, values for range parameters from the URL were placed
without any units added. Placing a value without a unit implies that
this value is given with a default unit multipler. However, the value
from the URL is represented in SI unit, and the default multiplier is
usually represented in a different unit.