Add category preloader and fix table display
This commit adds a loading gif while waiting for the elements of a
certain category to load. Also, table display has some fixes. A top
position of the table head is adjusted to new styles so that it sticks
to the top without extra space. Before this commit, some categories
seemed to skip some of their columns. That occured because a new value
type "o_t_val" wasn't considered while composing a JSON file with
elements and their info. To fix that, a simple elif statement was added
with the same body as in the "o_val" case.