Move name search field to filters
In this commit, the input field for searching by element name is moved
to the rest of the filters. Searching by name is working now and
some lines of code are put into functions, such as "filterPreloader" and
"composeURL".

In "displayElements" function, inner html of the table is rewritten
instead of being appended, and the limit variable is set to 200, instead
of being dynamically changed depending on the current number of
displayed elements. It is done because pages for displaying elements
will be introduced in the future.