Fix paging and unnecessary code fix
Before this commit, hrefTextPrefix in "updatePaginator" was set to an
empty string. Because of that, clicking on the page number changed the
URL from [host]/category/<category_id> to [host]/category/<page_number>.
This means that clicking on the page number 2 redirected to the category
number 2. Now the page number is appended to the end of the category
after the hash sign.

Because the currentPage variable is set to retrievePageFromURL(),
entering the URL with a specified page number actually changes the page
number. Without that, page were set to the first one everytime the URL
is entered and that caused trouble when pressing "back" and "forward"
buttons in the browser.