Fix data fetch when using back and forward buttons
Before this commit, the data was cached by the browser. Every fetch of
a category without filters was done with fetch('/catget'). Browser
caches data if the request is the same. Because of this, going to
another category in the breadcrumbs and going back by pressing the back
arrow button in the browser caused the data to be received from cache
and by the fetch function.